requirements.txt
-i https://pypi.org/simple
amqp==5.3.1; python_version >= '3.6'
asgiref==3.10.0; python_version >= '3.9'
billiard==4.2.1; python_version >= '3.7'
boto3==1.40.55; python_version >= '3.9'
botocore==1.40.55; python_version >= '3.9'
celery==5.4.0; python_version >= '3.8'
certifi==2025.1.31; python_version >= '3.6'
cffi==2.0.0; python_version >= '3.9'
charset-normalizer==3.4.1; python_version >= '3.7'
click==8.1.8; python_version >= '3.7'
click-didyoumean==0.3.1; python_full_version >= '3.6.2'
click-plugins==1.1.1
click-repl==0.3.0; python_version >= '3.6'
cryptography==46.0.2; python_version >= '3.8' and python_full_version not in '3.9.0, 3.9.1'
defusedxml==0.8.0rc2; python_version >= '3.6'
dj-database-url==2.3.0
django==5.2.7; python_version >= '3.10'
django-cors-headers==4.6.0; python_version >= '3.9'
django-debug-toolbar==4.4.6; python_version >= '3.8'
django-filter==24.3; python_version >= '3.8'
django-otp==1.5.4; python_version >= '3.7'
django-redis==5.4.0; python_version >= '3.6'
django-sendgrid-v5==1.3.0
django-silk==5.3.2; python_version >= '3.9'
django-templated-mail==1.1.1
djangorestframework==3.16.1; python_version >= '3.9'
djangorestframework-simplejwt==5.3.1; python_version >= '3.8'
djoser==2.3.1; python_version >= '3.8' and python_version < '4.0'
drf-nested-routers==0.94.1; python_version >= '3.8'
flower==2.0.1; python_version >= '3.7'
gprof2dot==2024.6.6; python_version >= '3.8'
gunicorn==23.0.0; python_version >= '3.7'
humanize==4.11.0; python_version >= '3.9'
idna==3.10; python_version >= '3.6'
jmespath==1.0.1; python_version >= '3.7'
kombu==5.4.2; python_version >= '3.8'
markupsafe==3.0.3; python_version >= '3.9'
oauthlib==3.2.2; python_version >= '3.6'
packaging==24.2; python_version >= '3.8'
pillow==11.0.0; python_version >= '3.9'
prometheus-client==0.21.1; python_version >= '3.8'
prompt-toolkit==3.0.48; python_full_version >= '3.7.0'
psycopg==3.2.3; python_version >= '3.8'
pycparser==2.23; python_version >= '3.8'
pyjwt==2.10.1; python_version >= '3.9'
pyotp==2.9.0; python_version >= '3.7'
python-dateutil==2.9.0.post0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2'
python-http-client==3.3.7; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
python3-openid==3.2.0
pytz==2024.2
razorpay==1.4.2
redis==5.2.1; python_version >= '3.8'
requests==2.32.3; python_version >= '3.8'
requests-oauthlib==2.0.0; python_version >= '3.4'
s3transfer==0.14.0; python_version >= '3.9'
sendgrid==6.12.5; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
six==1.17.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2'
social-auth-app-django==5.4.2; python_version >= '3.8'
social-auth-core==4.5.4; python_version >= '3.8'
sqlparse==0.5.3; python_version >= '3.8'
tornado==6.4.2; python_version >= '3.8'
tzdata==2024.2; python_version >= '2'
urllib3==2.5.0; python_version >= '3.9'
vine==5.1.0; python_version >= '3.6'
wcwidth==0.2.13
werkzeug==3.1.3; python_version >= '3.9'
whitenoise==6.8.2; python_version >= '3.9'
